refactor: create initGame.js and place StartGame() and RestartGame() functions
